server/requirements.txt
aniso8601==9.0.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
Flask-RESTful==0.3.10
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
PyJWT==2.8.0
pymongo==4.7.3
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
Werkzeug==3.0.3
WTForms==3.1.2
